Breaking up overlay into rendering and compositing classes
